+++ src/browser/getPinoBrowserWriteLog.ts
@@ -1,30 +1,24 @@
 import { getPinoWriteLog } from '../core/getPinoWriteLog'
-import { LOG_VALUE_TO_LABEL } from '../core/toSplunkEvent'
+import { PrettyConsole } from '../core/PrettyConsole'
+import { getConsoleLevelFromSplunkLog } from '../core/ConsoleLevel'
 
-import type { PinoLevelLabel, SplunkLog } from '../core/toSplunkEvent'
+import type { ConsoleLevel } from '../core/ConsoleLevel'
+import type { SplunkLog } from '../core/toSplunkEvent'
 import type SplunkEvents from 'splunk-events'
 
-type ConsoleLevel = Exclude<PinoLevelLabel, 'fatal'>
-
 type BrowserWriteLogCallbacks = {
-  onConsoleLog?: (level: ConsoleLevel, log: SplunkLog) => void
-}
-
-const defaultLog: BrowserWriteLogCallbacks['onConsoleLog'] = (level, log) => {
-  // eslint-disable-next-line no-console
-  console[level](log)
+  onConsoleLog?: (log: SplunkLog, level: ConsoleLevel) => void
 }
 
 export const getPinoBrowserWriteLog = (
   splunk: SplunkEvents,
-  { onConsoleLog = defaultLog }: BrowserWriteLogCallbacks = {}
+  { onConsoleLog = PrettyConsole.print }: BrowserWriteLogCallbacks = {}
 ) => {
   return getPinoWriteLog(splunk, {
     onSplunkLog: (log) => {
-      const levelLabel = LOG_VALUE_TO_LABEL[log.level]
-      const consoleLevel = levelLabel === 'fatal' ? 'error' : levelLabel
+      const consoleLevel = getConsoleLevelFromSplunkLog(log)
 
-      onConsoleLog(consoleLevel, log)
+      onConsoleLog(log, consoleLevel)
     },
   })
 }