+++ src/core/utils/logic/isPresent.ts
@@ -1,0 +1,3 @@
+export const isPresent = (record: Record<any, unknown>) => {
+  return Object.keys(record).length !== 0
+}